Removed an old workaround that creates problems now
Years ago, a workaround had been added for old versions of Qt5. However, not only it is not needed anymore but also it interferes with pressing `Enter` (messes with `Enter` after `Arrow`). So, it is reversed by this patch.

If someone can compile lxqt-config with an old version of Qt, they will see nothing special because the workaround didn't work well in old times.