Apply GTK settings only if the user chooses that
Fixes https://github.com/lxqt/lxqt-config/issues/508

Changing GTK settings is optional in the GUI; so, they shouldn't be overwritten if the user doesn't want that.

As to why changing GTK settings should remain optional, suffice it to say that the user may have a GTK-based DE installed alongside LXQt, although that isn't the only reason.