Fixed the "Set Position" page in lxqt-config-monitor (#1081)
The latest version of Qt 6.8 showed a problem in the code: the "Set Position" page was visible behind the "Fast Menu" page on startup, although everything worked fine.

The reason was that `MonitorPictureDialog` was a `QDialog` that was added to a stacked widget. The problem was solved by changing it to a `QWidget`. However, I didn't change its name to `MonitorPictureWidget`, because I didn't want to make a messy patch.

Also, a memory leak is fixed inside the same class.