Fixed clicking rectangles in Configuration Center (#1086)
Previously, depending on the window size and the arrangement of items, the lower parts of some clicking rectangles might be removed, resulting in no action when those lower parts were clicked.

The reason was that the logic of what is called "binary search" in the code works fine only if the grid rectangles are considered, not the item rectangles.

NOTE: At first I wanted to replace "binary search" with an ordinary search because it worked only with a grid. However, since we do set a grid in the code, I've kept its beautiful logic.