Added monospace setting to lxqt-config-appearance
A good monospace is needed by Qt for use in terminal emulators and code editors.

Also see https://github.com/lxqt/qterminal/wiki#how-to-set-my-preferred-font-for-monospace for how to set it up manually. LXQt users won't need to do that after this patch; the GUI will do it for them (in another way).